solution in java
This solution is passing all the test cases.
The logic behind the solution is, first of all we sort the object of the type medicine in ascending order of their associated price using bobble sort.
Doing so we will get the object array of type Medicine which will be sorted in ascending order of their associated price.
then we will find the disease that is coming as a parameter inside the function named medicinePriceForGivenDisease(). 
we will search for all the objects having the same disease name that is the passed parameter have.
if the disease name is matching with the disease name of the object's disease name then we will add the price of that medicine object into the result array of integer type and at last we will return the result array.  which will contain the price in ascending order of the given disease.